child.cpp: validate_cmd: Support symbolic links in environment variables
On some systems /bin is a symbolic link to /usr/bin (/bin -> /usr/bin), When
using the '-c' parameter to specify the executable program in the environment
variable, the problem of "multiple binary files matching" occurs, for example:

  $ sudo bpftrace -e 'kprobe:do_nanosleep { printf("%d sleeping\n", pid); }' -c 'sleep 5'
  ERROR: Failed to fork child: path 'sleep' must refer to a unique binary but matched 2 binaries

Debugging the bpftrace program through gdb can also be verified:

  (gdb) p valid_executable_paths
  $9 = std::vector of length 2, capacity 2 = {"/bin/ls", "/usr/bin/ls"}

This commit fixes the issue. The v1[0] will cause CICD busybox error.

[0] https://github.com/iovisor/bpftrace/pull/2689

Signed-off-by: Rong Tao <rongtao@cestc.cn>