codegen: IR: Fix ICmp instruction are not of the same type
Since commit 4ccf8912cf51 ("Treat SizedType ptr types as ptrs in codegen
(#4449)"), `Predicate` only checks the IsBoolTy() type, but this is not enough
for strncmp() UInt64.

We just change semantic_analyzer to make the return type a bool instead of a
CreateUInt64, and it's a breaking change.

Error example: (llvm-20.1.8)

    $ cat a.bt
    tracepoint:syscalls:sys_enter_open,
    tracepoint:syscalls:sys_enter_openat,
    tracepoint:syscalls:sys_enter_openat2
    / !strncmp(comm, "open", 4) /
    {
        printf("%s\n", comm);
    }

    Assertion:

    $ sudo a.bt
    bpftrace: /usr/include/llvm/IR/Instructions.h:1162: void llvm::ICmpInst::AssertOK(): Assertion `getOperand(0)->getType() == getOperand(1)->getType() && "Both operands to ICmp instruction are not of the same type!"' failed.
    Aborted

    The scoped_expr is 'i1', but the 'strncmp()' Constant::getNullValue is 'i64'.

Fix: commit 4ccf8912cf51 ("Treat SizedType ptr types as ptrs in codegen (#4449)")

Signed-off-by: Rong Tao <rongtao@cestc.cn>