kfunc: Check whether the BPF program type is supported
A large number of kfuncs have been introduced into the kernel, and these
kfuncs have different support for BPF program types across different kernel
versions. This commit adds a built-in function __builtin_kfunc_avail(),
which is used to determine at compile time whether a certain kfunc is
supported in the current program type.

For example, kfunc bpf_task_from_pid() was introduced in linux v6.1, its
support for tracepoint is in v6.12 [1]. We support comm(pid) in [2], but
there is a problem: when calling comm(pid) in kprobe, the verifier will
complain.

    tracepoint:syscalls:sys_enter_kill,
    tracepoint:syscalls:sys_enter_tkill,
    tracepoint:syscalls:sys_enter_tgkill,
    kprobe:__x64_sys_kill,
    kprobe:__x64_sys_tgkill,
    kprobe:__x64_sys_tkill
    {
      printf("kprobe %s\n", comm(pid));
    }

    ;  @ task.bpf.c:13
    28: (bf) r1 = r0
    29: (85) call bpf_task_from_pid#89392
    calling kernel function bpf_task_from_pid is not allowed

Therefore, we can optimize the script as follows:

    {
      if comptime  __builtin_kfunc_avail("bpf_task_from_pid") {
        printf("comm %s from %s\n", comm(pid), probe);
      } else {
        warnf("not support comm(pid) in %s\n", probe);
      }
    }

[1] https://docs.ebpf.io/linux/kfuncs/bpf_task_from_pid/
[2] commit 57dec2a8480a ("stdlib: comm(): support PID argument (#4799)")

Signed-off-by: Rong Tao <rongtao@cestc.cn>