Respect `missing_probes` config for bad `args` if the probe will be pruned
There is an interesting bug: when `config::missing_probes=warn`, if a
non-existent tracepoint is used and the built-in variable `args` is called,
`args` will be parsed for the non-existent tracepoint, resulting in an error.

For example, `tools/opensnoop.bt` on an aarch64 environment will throw an
error because the `open(2)` system call does not exist. This is clearly
incorrect.

    $ uname -m
    aarch64

    $ sudo bpftrace opensnoop.bt
    opensnoop.bt:49:1-35: ERROR: Tracepoint not found: syscalls:sys_enter_open
    tracepoint:syscalls:sys_enter_open,
    ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
    opensnoop.bt:130:1-34: ERROR: Tracepoint not found: syscalls:sys_exit_open
    tracepoint:syscalls:sys_exit_open,
    ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

Simplified analysis example (on aarch64):

  Works fine:
    config = { missing_probes = warn; }
    tracepoint:syscalls:sys_enter_open,
    tracepoint:syscalls:sys_enter_openat,
    tracepoint:syscalls:sys_enter_openat2
    {}
    WARNING: Unable to attach probe: tracepoint:syscalls:sys_enter_open. Skipping.

  Cannot work:
    config = { missing_probes = warn; }
    tracepoint:syscalls:sys_enter_open,
    tracepoint:syscalls:sys_enter_openat,
    tracepoint:syscalls:sys_enter_openat2
    {
      @filename[tid] = args.filename;
    }
    ./tp.bt:7:1-35: ERROR: Tracepoint not found: syscalls:sys_enter_open
    tracepoint:syscalls:sys_enter_open,
    ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

This submission properly filters out attachments of type ExpansionType::NONE
in ProbeAndApExpander in the case of `config::missing_probes!=error`.

This also requires modifications to the test cases. The attachpoints used
in the test cases need to be defined in gmock. For example, gmock's
get_symbols_from_traceable_funcs() must return all functions used throughout
the unit tests, otherwise attach points will be skipped in `ProbeAndApExpander`
which calls `get_symbols_from_traceable_funcs(false)`.

Signed-off-by: Rong Tao <rongtao@cestc.cn>